Fix URI template operators 'prefix', 'suffix' and 'list' to return the empty string for undefined variables.
As per draft 3 of the URI template spec, made the the 'prefix', 'suffix' and 'list' to return the empty string for undefined variables.
Previously, 'prefix' and 'suffix' returned the arg regardless and 'list' raised an exception.

Added tests.  Ideally the next draft of the spec would include examples equivalent to these.