avoid producing invalid urls for partial form-style query expansion
`http://example.com/{?a,b}` partially expanded with `b: 2` must result in
`http://example.com/?b=2{&a}`. That way if the final expansion done w/ `a`
undefined the resulting URL will be valid. If the form-style query vars are not
reordered and final expansion is done w/ `a` undefined the result is
`http://example.com/&b=2` which is clearly malformed.
remove superfluous code in operator sequence enumerator
fix hang issue on rbx