Only include the key files in the gem artifact
This change ships just the necessary files for the gem + the license,
changelog and readme. This avoids shipping things like the specs and
rake tasks. It cuts the size of the gem artifact by 25%+ and reduces
file count on disk significantly which improves performance on windows
hosts as they have to AV scan each file written.

Signed-off-by: Tim Smith <tsmith@chef.io>