Only translate plus to space for http schemes
Prompted by a [discussion on twitter][].

Prior to this commit query_values replaced "+" with a space regardless
of the scheme. This makes sense for http and https, but it might not
make sense for other schemes.

For example, in the case of a mailto like
`mailto:?to=example+to@example.com&cc=example+cc@example.com`, a space
is probably not what we want, since that would not be a valid email. I
don't see explicit guidance on this in [RFC 6068][], although there
is this:

[RFC 6068]: http://www.faqs.org/rfcs/rfc6068.html

> Software creating 'mailto' URIs likewise has to be careful to encode
> any reserved characters that are used.  HTML forms are one kind of
> software that creates 'mailto' URIs.  Current implementations encode
> a space as '+', but this creates problems because such a '+' standing
> for a space cannot be distinguished from a real '+' in a 'mailto'
> URI.  When producing 'mailto' URIs, all spaces SHOULD be encoded as
> %20, and '+' characters MAY be encoded as %2B.  Please note that '+'
> characters are frequently used as part of an email address to
> indicate a subaddress, as for example in <bill+ietf@example.org>.

This commit changes the behavior of `Addressable::URI#query_values` to
only replace plus with space when the scheme is http, https or nil.

[discussion on twitter]: https://twitter.com/sporkmonger/status/1346239980717314048