Replace the `unicode.data` blob by a ruby constant
Ref: 1eb715d16c93987f5feb5507ad88f6a7cd708b61

It seems like this data was initially YAML, but 13 years ago
various version of libyaml had issues, so Marshal was used.

The problem with marshal is it's a binary format, so while this
data has no reason to change, it's hard to see what's in it.

I think a simple Ruby constant would work just as well.
On my machine, both take as long to load, but many apps have
bootsnap setup, so would probably load the ruby version faster
than the marshal one.